Brandon_CMT_FishPullLevers
Sorta, They do the same thing as the lever at the moment and when an
animation is added I will update it, it works as it should for gameplay
purposes and can also toggle lights. What is does is changable by the
tag it has, "Anchovies" tag turns off where as "AntiChovies" Turns
things on.